Moved source reader to renderer namespace
