+++ src/Renderer/SourceReaderInterface.php
@@ -1,5 +1,5 @@
 <?php
-namespace ZeroConfig\Preacher\Generator;
+namespace ZeroConfig\Preacher\Renderer;
 
 use ZeroConfig\Preacher\Source\SourceInterface;
